+++ src/main/java/com/adoonge/seedzip/member/domain/MemberAiUsages.java
@@ -1,0 +1,40 @@
+package com.adoonge.seedzip.member.domain;
+
+import java.time.LocalDate;
+
+import com.adoonge.seedzip.global.entity.BaseEntity;
+
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.OneToMany;
+import jakarta.persistence.OneToOne;
+import jakarta.persistence.Table;
+import lombok.AccessLevel;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@Builder
+@Entity
+@Table(name = "member_ai_usages")
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+@AllArgsConstructor
+public class MemberAiUsages extends BaseEntity {
+
+	@Id
+	@GeneratedValue(strategy = GenerationType.IDENTITY)
+	private Long id;
+
+	@OneToOne
+	@JoinColumn(name = "member_id")
+	private Member memberId;
+
+	private LocalDate lastUsedDate;
+
+	private Long usageCount;
+}
